Formatting

+++ Clutch/ZipArchive.m
@@ -93,13 +93,13 @@
     return [self CreateZipFile2:zipFile append:isAppend];
 }
 
-    /**
-     * add an existing file on disk to the zip archive, compressing it.
-     *
-     * @param file    the path to the file to compress
-     * @param newname the name of the file in the zip archive, ie: path relative to the zip archive root.
-     * @returns BOOL YES on success
-     */
+/**
+ * add an existing file on disk to the zip archive, compressing it.
+ *
+ * @param file    the path to the file to compress
+ * @param newname the name of the file in the zip archive, ie: path relative to the zip archive root.
+ * @returns BOOL YES on success
+ */
 
 #define M_FRAGMENT_SIZE (1024 * 1024 * 15)
 
@@ -158,7 +158,7 @@
                                   0,
                                   NULL, // comment
                                   Z_DEFLATED,
-                                  self.compression);
+                                  (int)self.compression);
     } else {
         FILE *f = fopen([file cStringUsingEncoding:NSUTF8StringEncoding], "r");
         fseek(f, 0, SEEK_END);
@@ -178,7 +178,7 @@
                                    0,
                                    NULL, // comment
                                    Z_DEFLATED,
-                                   self.compression,
+                                   (int)self.compression,
                                    0,
                                    15,
                                    8,
@@ -277,7 +277,7 @@
                                   0,
                                   NULL, // comment
                                   Z_DEFLATED,
-                                  self.compression);
+                                  (int)self.compression);
     } else {
         uLong crcValue = crc32(0L, NULL, 0L);
         crcValue = crc32(crcValue, (const Bytef *)data.bytes, (unsigned int)data.length);
@@ -290,7 +290,7 @@
                                    0,
                                    NULL, // comment
                                    Z_DEFLATED,
-                                   self.compression,
+                                   (int)self.compression,
                                    0,
                                    15,
                                    8,
